feat: Add ability to open/close Device Panel as needed
With buttplug-js 0.6.0, we no longer have a singleton
TestDeviceManager, so we have to extract the server and send it. Even
then, that may fail if the server doesn't have a TestDeviceManager.
Update the display macro to log to console on failure, and create a
close macro so we can close the panel on disconnect (as the panel will
be invalid after disconnect anyways).